Implements two differents types of list for ListProperty
- Implement ArrayList
- Implement LinkedList
