+++ src/main/java/fr/ourten/teabeans/value/LinkedListProperty.java
@@ -1,0 +1,25 @@
+package fr.ourten.teabeans.value;
+
+import java.util.List;
+
+import com.google.common.collect.Lists;
+
+/**
+ * Implementation of ListProperty using LinkedList
+ * 
+ * @author Phenix246
+ */
+public class LinkedListProperty<T> extends ListProperty<T>
+{
+
+    public LinkedListProperty(final List<T> value, final String name)
+    {
+        super(value != null ? Lists.newLinkedList(value) : Lists.newLinkedList(), name);
+    }
+
+    public LinkedListProperty(final List<T> value)
+    {
+        this(value, "");
+    }
+
+}
